pyproject.toml
[project]
name = "events-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "dakshmakkar6-code", email = "dakshmakkar333@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.0",
    "cryptography>=45.0.7",
    "dateparser>=1.2.2",
    "discord-py>=2.6.3",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "lxml>=6.0.0",
    "openai>=1.93.0",
    "prometheus-client>=0.22.1",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pytz>=2024.1",
    "requests>=2.31.0",
    "rich>=14.0.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "structlog>=25.4.0",
    "supabase>=2.0.0",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.35.0",
    "yfinance>=0.2.64",
]

[project.scripts]
events-agent = "events_agent.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["events_agent*"]
